Move away from hard-coded machine name
Adding more architecture types or moving machine around was starting to
take its toll on how often we reconfigure the jenkins scripts. This
change decouples which physical machine actually runs the jobs and makes
sure that only a provisioner change (machine name) is needed to relocate
resources to the right place.