Remove all kernels, install reference
This is necessary as, if the OBS repo is down, only the upstream kernel
is installed in the system and the next reboot fails with widly
different error messages on different architectures, making it a living
hell to debug what is going on.

Removing the kernel with OBS means we *only* get the reference kernel
when OBS is up and we get a BIOS screen when OBS is down, because the
only entry is "System Setup", exchanging random behaviour for identical
behaviour on all machines.