Bazel: Build RPM or Install it
This commit overhauls Bazel's build by adding a spec file and building
from that.
Note that since we're building Bazel 0.26.1 for TF 1.15.0, the upstream
Fedora spec file has to be patched to drop the bash-completion target.
That said I have built Bazel 1.1.0 (which can be used for TF 2 git)
using the upstream spec file, it is just that bash-completion requires
>0.27.1
Also, java has been bumped from 8 to 11.